Move project to src layout
