+++ src/bulletin_builder/app_core/loader.py
@@ -1,14 +1,16 @@
 import importlib
 import pkgutil
-import app_core
+
 
 def init_app(app):
     """
     Initialize core attributes and then discover all feature modules under app_core,
     calling their `init(app)` functions to register functionality.
     """
+    base_pkg = __name__  # e.g., 'bulletin_builder.app_core'
+
     try:
-        from app_core.core_init import init as _core_init
+        from .core_init import init as _core_init
         _core_init(app)
     except Exception as e:
         print(f"Error in core_init: {e}")
@@ -16,7 +18,7 @@ def init_app(app):
     # 2) then drafts, handlers, sections, exporter, preview...
     for module in ("handlers", "drafts", "sections", "exporter", "preview", "ui_setup"):
         try:
-            m = importlib.import_module(f"app_core.{module}")
+            m = importlib.import_module(f"{base_pkg}.{module}")
             if hasattr(m, "init"):
                 m.init(app)
         except Exception as e:
