+++ src/bulletin_builder/bulletin_renderer.py
