+++ src/bulletin_builder/templates/partials/announcements.html
