+++ src/bulletin_builder/templates/partials/custom_text.html
