+++ src/bulletin_builder/templates/partials/default.html
