+++ src/bulletin_builder/templates/partials/events.html
