+++ src/bulletin_builder/templates/partials/image.html
