+++ src/bulletin_builder/templates/themes/halloween.css
