+++ src/bulletin_builder/templates/themes/thanksgiving.css
