+++ src/bulletin_builder/ui/base_section.py
