+++ src/bulletin_builder/ui/custom_text.py
@@ -1,5 +1,5 @@
 import customtkinter as ctk
-from ui.base_section import SectionRegistry
+from .base_section import SectionRegistry
 
 @SectionRegistry.register("custom_text")
 class CustomTextFrame(ctk.CTkFrame):
